# api/openapi.yaml
...
openapi: 3.0.3
info:
  title: MLAKP Shared Expense API
  version: 0.1.0
  description: HTTP API contract for the shared expense and debt tracking backend.
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/auth/register:
    post:
      tags:
        - Auth
      summary: Register a user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              name: Thomas
              username: thomas
              email: thomas@example.com
              password: password123
      responses:
        "201":
          description: User registered and access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
              example:
                success: true
                data:
                  access_token: eyJzdWIiOiIxMjMifQ.signature
                  refresh_token: opaque-refresh-token
                  token_type: Bearer
                  expires_at: "2026-05-02T12:15:00Z"
                  user:
                    id: 3b43c6da-2f6f-4f46-ae15-f2f6a25dfe62
                    name: Thomas
                    username: thomas
                    email: thomas@example.com
        "400":
          description: Invalid JSON or invalid user input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidJson:
                  value:
                    success: false
                    error:
                      code: invalid_json
                      message: Request body must be valid JSON
                invalidName:
                  value:
                    success: false
                    error:
                      code: invalid_name
                      message: Name must be between 1 and 120 characters
                invalidEmail:
                  value:
                    success: false
                    error:
                      code: invalid_email
                      message: Email is invalid
                invalidUsername:
                  value:
                    success: false
                    error:
                      code: invalid_username
                      message: Username must be 3 to 30 lowercase letters, numbers, or underscores
                invalidPassword:
                  value:
                    success: false
                    error:
                      code: invalid_password
                      message: Password must be at least 8 characters
        "409":
          description: Email or username is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: email_already_registered
                  message: Email is already registered
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/login:
    post:
      tags:
        - Auth
      summary: Login a user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              email: thomas@example.com
              password: password123
      responses:
        "200":
          description: Credentials accepted and access token issued.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
              example:
                success: true
                data:
                  access_token: eyJzdWIiOiIxMjMifQ.signature
                  refresh_token: opaque-refresh-token
                  token_type: Bearer
                  expires_at: "2026-05-02T12:15:00Z"
                  user:
                    id: 3b43c6da-2f6f-4f46-ae15-f2f6a25dfe62
                    name: Thomas
                    username: thomas
                    email: thomas@example.com
        "400":
          description: Invalid JSON or invalid email syntax.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidJson:
                  value:
                    success: false
                    error:
                      code: invalid_json
                      message: Request body must be valid JSON
                invalidEmail:
                  value:
                    success: false
                    error:
                      code: invalid_email
                      message: Email is invalid
        "401":
          description: Credentials are invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: invalid_credentials
                  message: Email or password is incorrect
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/logout:
    post:
      tags:
        - Auth
      summary: Logout current client
      description: Revokes the current server-side session. Access and refresh tokens tied to the session are rejected after logout.
      operationId: logoutUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logout accepted and current session revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data: null
        "401":
          description: Missing, malformed, expired, invalid, or stale access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: invalid_access_token
                  message: Access token is invalid or expired
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
            example:
              refresh_token: opaque-refresh-token
      responses:
        "200":
          description: Refresh token accepted and rotated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshTokenResponse"
              example:
                success: true
                data:
                  access_token: eyJzdWIiOiIxMjMifQ.signature
                  refresh_token: rotated-refresh-token
                  token_type: Bearer
                  expires_at: "2026-05-02T12:15:00Z"
        "400":
          description: Invalid JSON.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: invalid_json
                  message: Request body must be valid JSON
        "401":
          description: Refresh token is invalid, expired, rotated, or revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                success: false
                error:
                  code: invalid_refresh_token
                  message: Refresh token is invalid or expired
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/send-otp:
    post:
      tags:
        - Auth
      summary: Send OTP to email
      description: Sends a 6-digit OTP to the specified email for verification or password reset.
      operationId: sendOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendOTPRequest"
            example:
              email: user@example.com
              purpose: signup
      responses:
        "200":
          description: OTP sent successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  message: OTP sent successfully
                  expires_at: "2026-05-02T12:25:00Z"
                  purpose: signup
        "400":
          description: Invalid email or purpose.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered (for signup purpose).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          description: Rate limit exceeded or cooldown active.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/verify-otp:
    post:
      tags:
        - Auth
      summary: Verify OTP
      description: Verifies the signup OTP and completes email verification. Use /v1/auth/reset-password for password reset OTPs.
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyOTPRequest"
            example:
              email: user@example.com
              otp: "123456"
              purpose: signup
      responses:
        "200":
          description: OTP verified successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Invalid or expired OTP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/reset-password:
    post:
      tags:
        - Auth
      summary: Reset password with OTP
      description: Verifies OTP and resets the user's password. Revokes all existing sessions.
      operationId: resetPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
            example:
              email: user@example.com
              otp: "123456"
              new_password: newpassword123
      responses:
        "200":
          description: Password reset successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  message: Password reset successfully
        "400":
          description: Invalid request or OTP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/auth/send-otp/account:
    post:
      tags:
        - Auth
      summary: Send OTP to authenticated user's email
      description: Sends a password reset OTP to the currently authenticated user's email.
      operationId: sendOTPForAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OTP sent successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              example:
                success: true
                data:
                  message: OTP sent successfully
                  expires_at: "2026-05-02T12:25:00Z"
        "401":
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "429":
          description: Rate limit exceeded or cooldown active.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/me:
    get:
      tags:
        - Users
      summary: Get current authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
              example:
                success: true
                data:
                  user:
                    id: 3b43c6da-2f6f-4f46-ae15-f2f6a25dfe62
                    name: Thomas
                    username: thomas
                    email: thomas@example.com
        "401":
          description: Missing, malformed, expired, invalid, or stale access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                unauthenticated:
                  value:
                    success: false
                    error:
                      code: unauthenticated
                      message: Authentication is required
                invalidAuthorizationHeader:
                  value:
                    success: false
                    error:
                      code: invalid_authorization_header
                      message: Authorization header must use Bearer token
                invalidAccessToken:
                  value:
                    success: false
                    error:
                      code: invalid_access_token
                      message: Access token is invalid or expired
                missingUser:
                  value:
                    success: false
                    error:
                      code: invalid_access_token
                      message: Access token user no longer exists
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags:
        - Users
      summary: Update current authenticated user
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRequest"
            example:
              username: thomas_new
      responses:
        "200":
          description: Updated current user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserResponse"
        "400":
          description: Invalid username.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing, malformed, expired, invalid, or stale access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username is already registered.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/users/search:
    get:
      tags:
        - Users
      summary: Search users by username prefix
      operationId: searchUsers
      security:
        - bearerAuth: []
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            pattern: "^[a-z0-9_]+$"
          description: Case-insensitive username prefix using lowercase letters, numbers, or underscores. Returns up to 10 users.
      responses:
        "200":
          description: Matching users.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserSearchResponse"
        "401":
          description: Missing, malformed, expired, invalid, or stale access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/groups:
    post:
      tags:
        - Groups
      summary: Create a group
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGroupRequest"
            example:
              name: Home
      responses:
        "201":
          description: Group created. The creator is added as owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupResponse"
              example:
                success: true
                data:
                  group:
                    id: b2c1c30f-5c3e-4376-9d4e-ef56ef7c8b81
                    name: Home
                    created_by: 3b43c6da-2f6f-4f46-ae15-f2f6a25dfe62
                    created_at: "2026-05-03T10:15:00Z"
                    updated_at: "2026-05-03T10:15:00Z"
        "400":
          description: Invalid JSON or invalid group input.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidJson:
                  value:
                    success: false
                    error:
                      code: invalid_json
                      message: Request body must be valid JSON
                invalidGroupName:
                  value:
                    success: false
                    error:
                      code: invalid_group_name
                      message: Group name must be between 1 and 120 characters
        "401":
          description: Missing, malformed, expired, invalid, or stale access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          $ref: "#/components/responses/RequestBodyTooLarge"
        "500":
          $ref: "#/components/responses/InternalError"
    get:
      tags:
        - Groups
      summary: List current user's groups
      operationId: listGroups
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Groups the current user belongs to.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupListResponse"
              example:
                success: true
                data:
                  groups:
                    - id: b2c1c30f-5c3e-4376-9d4e-ef56ef7c8b81
                      name: Home
                      created_by: 3b43c6da-2f6f-4f46-ae15-f2f6a25dfe62
                      created_at: "2026-05-03T10:15:00Z"
                      updated_at: "2026-05-03T10:15:00Z"
        "401":
          description: Missing, malformed, expired, invalid, or stale access token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/expenses/{expenseID}:
    get:
      tags:
        - Expenses
      summary: Get an expense
      description: Returns an expense with its participants and generated debts when the authenticated user is a group member.
      operationId: getExpense
      security:
        - bearerAuth: []
      parameters:
        - name: expenseID
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Expense details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExpenseDetailsResponse"
        "400":
          description: Invalid expense ID.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Expense not found or not visible to the current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: HMAC-SHA256
...
  schemas:
...
    SuccessResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          nullable: true
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum:
            - false
        error:
          $ref: "#/components/schemas/ErrorDetail"
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - username
        - email
        - password
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: "^[a-z0-9_]+$"
          description: Unique lowercase username used for member search.
        email:
          type: string
          format: email
          maxLength: 254
          description: Email is normalized to lowercase before storage.
        password:
          type: string
          format: password
          minLength: 8
          writeOnly: true
...
    AuthTokenResponse:
      type: object
      additionalProperties: false
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum:
            - true
        data:
          $ref: "#/components/schemas/AuthTokenData"
...